Reorganize assets/js with client/ and __tests__/ structure
Restructures JavaScript source and test files to improve organization,
maintainability, and align with package structure.

Structure:
- client/: Production code matching @arizona-framework/client package
  - arizona.js, arizona-hierarchical.js, arizona-worker.js
  - logger/: Logger subsystem (arizona-logger.js, arizona-console-logger.js)
  - index.js: Package entry point

- __tests__/: All testing code isolated from production
  - unit/: Unit tests (Vitest) mirroring client/ structure
  - e2e/: End-to-end tests (Playwright) organized by execution type
    - parallel/: Tests that can run concurrently
    - sequential/: Tests requiring sequential execution
    - utils.js: Shared e2e test utilities
  - __mocks__/: Test mocks (WebSocket, Worker)

- index.js: Root convenience re-export from client/

Changes:
- Move all source files to client/ directory
- Move all test files to __tests__/ with proper organization
- Update vite.config.js entry points and DTS includes
- Update vitest.config.js test patterns and coverage paths
- Update playwright.config.js test directories
- Update package.json npm scripts for e2e tests
- Update all import paths in source and test files
- Rename test-utils.js → utils.js for consistency

Benefits:
- Clear separation: production (client/) vs testing (__tests__/)
- Package alignment: client/ matches @arizona-framework/client
- Test type clarity: unit vs e2e vs mocks clearly separated
- Scalability: Room for future packages (server/, devtools/)
- Build simplicity: Easy to exclude __tests__/ from builds
- Industry standard: Follows common __tests__/ convention

Verification:
- ✅ 154/154 unit tests passing
- ✅ 101/102 e2e tests passing (1 pre-existing flaky Firefox test)
- ✅ Build successful with all bundles generated
- ✅ All git renames tracked properly
