+++ assets/js/__tests__/e2e/parallel/counter-app.spec.js
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { waitForCondition, collectWebSocketMessages } from '../test-utils.js';
+import { waitForCondition, collectWebSocketMessages } from '../utils.js';
 
 test.describe('Arizona Counter App', () => {
   test('should load counter page and display initial state', async ({ page }) => {
