+++ assets/js/__tests__/e2e/parallel/todo-app.spec.js
@@ -8,7 +8,7 @@ import {
   waitForInputValue,
   selectors,
   createAssertions,
-} from '../test-utils.js';
+} from '../utils.js';
 
 test.describe('Arizona Todo App', () => {
   // Create assertions with expect function
