+++ assets/js/__tests__/e2e/sequential/counter-pubsub.spec.js
