+++ assets/js/client/index.js
@@ -1,0 +1,6 @@
+// Main entry point for Arizona Framework client package
+export { default } from './arizona.js';
+export { default as ArizonaClient } from './arizona.js';
+
+// Re-export all logger exports
+export * from './logger/index.js';
