+++ assets/js/client/logger/arizona-console-logger.js
