+++ assets/js/client/logger/index.js
