+++ assets/js/index.js
@@ -1,6 +1,2 @@
-// Main entry point for Arizona Framework client
-export { default } from './arizona.js';
-export { default as ArizonaClient } from './arizona.js';
-
-// Re-export all loggers
-export * from './logger/index.js';
+// Root entry point - convenience re-export from client package
+export * from './client/index.js';
