+++ package.json
@@ -77,8 +77,8 @@
     "test:unit": "vitest run",
     "test:unit:coverage": "vitest run --coverage",
     "test:e2e": "npm run test:e2e:parallel && npm run test:e2e:sequential",
-    "test:e2e:parallel": "playwright test --reporter=line assets/js/e2e/parallel",
-    "test:e2e:sequential": "playwright test --reporter=line assets/js/e2e/sequential --workers=1",
+    "test:e2e:parallel": "playwright test --reporter=line assets/js/__tests__/e2e/parallel",
+    "test:e2e:sequential": "playwright test --reporter=line assets/js/__tests__/e2e/sequential --workers=1",
     "test:e2e:report": "playwright test",
     "test:e2e:ui": "playwright test --ui",
     "ci": "npm run format:check && npm run lint:check && npm test",
