+++ vitest.config.js
@@ -2,13 +2,13 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['assets/**/*.test.js'],
+    include: ['assets/js/__tests__/unit/**/*.test.js'],
     exclude: ['_build/**', '.github/**'],
     coverage: {
       provider: 'v8',
       reportsDirectory: '_build/test/cover/js',
-      include: ['assets/**/*.js'],
-      exclude: ['assets/**/*.test.js', 'assets/**/__mocks__/**', 'assets/**/e2e/**'],
+      include: ['assets/js/client/**/*.js'],
+      exclude: ['assets/js/__tests__/**'],
       thresholds: {
         statements: 60,
         branches: 80,
